T #38 Moved BaseRepository to kernel abstractions
