+++ Huppy/Huppy.Infrastructure/Repositories/ReminderRepository.cs
@@ -1,6 +1,7 @@
 using Discord.Interactions;
 using Huppy.Core.Interfaces.IRepositories;
 using Huppy.Core.Models;
+using Huppy.Kernel.Abstraction;
 using Microsoft.EntityFrameworkCore;
 
 namespace Huppy.Infrastructure.Repositories
