+++ Huppy/Huppy.Infrastructure/Repositories/TicketRepository.cs
@@ -1,5 +1,6 @@
 using Huppy.Core.Interfaces.IRepositories;
 using Huppy.Core.Models;
+using Huppy.Kernel.Abstraction;
 
 namespace Huppy.Infrastructure.Repositories
 {
