+++ Huppy/Huppy.Infrastructure/Repositories/UserRepository.cs
@@ -1,5 +1,6 @@
 using Huppy.Core.Interfaces.IRepositories;
 using Huppy.Core.Models;
+using Huppy.Kernel.Abstraction;
 using Microsoft.EntityFrameworkCore;
 
 namespace Huppy.Infrastructure.Repositories
